requirements.txt
Pillow
pycryptodome
numpy
cryptography